Implement public bookings functionality
Updated the bookings view and controller to implement the ability to create new bookings. Active prayer slots are now available as a dropdown in the form. Also, adjusted the prayer vigil and prayer slot models to support this functionality. Additionally, added destruction of associated prayer vigils when an event is deleted.